[fixed] some of the coverity defects

+++ MMOCoreORB/src/server/zone/objects/creature/commands/HealHealthSelf1Command.h
@@ -54,7 +54,7 @@ public:
 		uint32 healthHealed = creature->healDamage(creature, CreatureAttribute::HEALTH, healAmount);
 
 
-		forceCostDeducted = MIN((healthHealed / 7), forceCost);
+		forceCostDeducted = MIN(((float)healthHealed / (float)7), forceCost);
 
 
 		// Send system message(s).
